// src/Types/Products.type.ts
export interface Product {
  product_id: string;
  product_name: string;
  product_desc: string;
}
// export const Products: Product = {
//   product_id: "",
//   product_name: "",
//   product_desc: "",
// };

// Pick :Nhặt ra
// export type Products = Pick<
//   Product,
//   "product_id" | "product_name" | "product_desc"
// >[];